Moving styling and page logic up into ODTV page
